Reuse status function in entity

+++ src/Shell/PruneShell.php
@@ -4,7 +4,6 @@ namespace App\Shell;
 use App\Model\Entity\Message;
 use App\Model\Table\MessagesTable;
 use Cake\Console\Shell;
-use DateTime;
 
 /**
  * @property MessagesTable Messages
@@ -19,13 +18,11 @@ class PruneShell extends Shell
 
     public function main()
     {
-        $query = $this->Messages->find('all', [
-            'conditions' => ['Messages.end_date <' < new DateTime('-2 days')],
-        ]);
+        $query = $this->Messages->find('all');
+
         /** @var Message[] $messages */
         $messages = $query->all();
-        debug($messages);
-        return;
+
         if ($query->isEmpty()) {
             $this->out('No outdated messages found.');
             return true;
@@ -34,6 +31,11 @@ class PruneShell extends Shell
         $success = true;
         $this->out('Deleting files...');
         foreach ($messages as $message) {
+            if ($message->status !== 'verlopen') {
+                continue;
+            }
+
+            $this->out(sprintf("Deleting ‘%s’ (%s)", $message->name, $message->end_date));
             if (!$this->Messages->delete($message)) {
                 $success = false;
                 $this->err(sprintf('Error deleting ‘%s’ (%s)', $message->name, $message->path));
